Lab 15: Add validation

+++ src/main/java/chirp/service/resources/UserResource.java
@@ -2,6 +2,7 @@ package chirp.service.resources;
 
 import java.net.URI;
 
+import javax.ws.rs.BadRequestException;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.FormParam;
 import javax.ws.rs.GET;
@@ -21,6 +22,8 @@ import chirp.model.UserRepository;
 @Path("/users")
 public class UserResource {
 
+	private static final String USERNAME_REGEX = "^[a-z0-9_-]{3,16}$";
+
 	private final Logger log = LoggerFactory.getLogger(getClass());
 
 	private UserRepository repository = UserRepository.getInstance();
@@ -28,6 +31,9 @@ public class UserResource {
 	@GET
 	@Path("/{username}")
 	public Response getUser(@PathParam("username") String username) {
+		if (!validate(username)) {
+			throw new BadRequestException("Invalid username: " + username);
+		}
 		User user = repository.getUser(username);
 		return Response.ok().entity(user.toString()).build();
 	}
@@ -35,6 +41,9 @@ public class UserResource {
 	@PUT
 	@Path("/{username}")
 	public Response createUser(@PathParam("username") String username, @FormParam("realname") String realname, @Context UriInfo uriInfo) {
+		if (!validate(username)) {
+			throw new BadRequestException("Invalid username: " + username);
+		}
 		repository.createUser(username, realname);
 		URI location = uriInfo.getAbsolutePathBuilder().build();
 		return Response.created(location).build();
@@ -43,8 +52,15 @@ public class UserResource {
 	@DELETE
 	@Path("/{username}")
 	public Response removeUser(@PathParam("username") String username) {
+		if (!validate(username)) {
+			throw new BadRequestException("Invalid username: " + username);
+		}
 		repository.deleteUser(username);
 		return Response.noContent().build();
+	}
+	
+	private boolean validate(String username) {
+		return username != null && username.matches(USERNAME_REGEX);
 	}
 
 }
